refactor: ApiTest 위치 및 명칭 변경
[개요]
- 사용자뿐만 아니라 다른 도메인 테스트에서 재사용을 위해 명칭과 클래스 위치를 수정

[변경사항]
- 클래스 위치 변경
- UserApiTest -> ApiTest로 변경

[결과]
- ApiTest 명칭 변경
