+++ src/test/java/com/soompyo/server/ApiTest.java
@@ -1,4 +1,4 @@
-package com.soompyo.server.user;
+package com.soompyo.server;
 
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
@@ -7,12 +7,9 @@ import java.lang.annotation.Target;
 
 import org.springframework.boot.test.context.SpringBootTest;
 
-import com.soompyo.server.FixtureConfiguration;
-import com.soompyo.server.ServerApplication;
-
 @Target(ElementType.TYPE)
 @Retention(RetentionPolicy.RUNTIME)
 @SpringBootTest(classes = {ServerApplication.class,
     FixtureConfiguration.class}, webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public @interface UserApiTest {
+public @interface ApiTest {
 }
